tests: Set VK_LAYER_PATH automatically
When tests are executing, they need to find the layers built by this
project. However, the only robust way to do that is to set the
VK_LAYER_PATH environment variable with the correct path. While it
would be simple to set the build path of the layers into a header file,
because of multi-configuration (like visual studio), some extra CMake
logic is required to rename the header file containing the path using
the current build configuration as the source.