layers: Fix EnumPhysDevs in decompression
EnumeratePhysicalDevices was only looking for VK_SUCCESS when trying to
update its own maps of found physical device handles. This is a problem
when the application intentionally only queries for 1 physical device,
getting VK_INCOMPLETE but not following up for all physical devices.
The tests intentionally look for 1 physical device, triggering this
error.

The fix is to update the map of found physical devices if either
VK_SUCCESS or VK_INCOMPLETE is returned.