deps: Update to latest Vulkan-Utility-Libraries
Move this repo foward to the latest Vulkan-Utility-Libraries commit.
This enables usage of the vk_struct_helper.hpp and removes code
that was responsible for generating vk_typemap_helper.h, which has the
same functionality.

This commit deletes layer_dispatch_table_generator.py as it was unused
and the same header file can now be found in Vulkan-Utility-Libraries.