tests: Make failed assertions throw
Add a listener to the test framework which throws an exceptions when a
test failure occurs. This makes it so that all failed assertions will
cause tests to stop executing and begin tearing down, rather than
allowing subsequent test code to run, which may put drivers in a bad
state.